insert and select work
`insert 0 f0c1s him@f0c1s.com pass` will enter a record in the database
This is in-memory still.
`select` will fetch the records in the database.

However you can enter same id and it will still insert with no care
in the world for id value.
`insert 0 a b c`
`insert 0 x y z`
`select`
// Will fetch both of the lines
[0, a, b, c]
[0, x, y, z]

So, we have a liftoff, but we might very well end up on the dark side of the Sun.
Sat Sep 30 02:51:22 AM IST 2023